+++ src/error.rs
@@ -1,5 +1,6 @@
-use rustc_serialize;
+use hexx;
 use curl;
+use serde_json;
 
 /// `Result` type-alias
 pub type Result<T> = ::std::result::Result<T, Error>;
@@ -20,18 +21,18 @@ quick_error! {
             display("utf8 error: {}", err)
             cause(err)
         }
-        /// `rustc_serialize::json::EncoderError`
-        Encoder(err: rustc_serialize::json::EncoderError) {
+        /// `serde_json::error::Error`
+        Serialize(err: serde_json::error::Error) {
             from()
-            description("rustc_serialize::json::EncoderError")
-            display("rustc_serialize::json::EncoderError: {}", err)
+            description("serde_json::error::Error")
+            display("serde_json::error::Error: {}", err)
             cause(err)
         }
         /// `rustc_serialize::hex::FromHexError`
-        FromHex(err: rustc_serialize::hex::FromHexError) {
+        FromHex(err: hexx::FromHexError) {
             from()
-            description("rustcrustc_serialize::hex::FromHexError")
-            display("rustc_serialize::hex::FromHexError: {}", err)
+            description("hexx::FromHexError")
+            display("hexx::FromHexError: {}", err)
             cause(err)
         }
         /// `HexColor` parsing error