+++ src/slack.rs
@@ -1,8 +1,8 @@
 use curl::easy::Easy;
 use std::str;
-use rustc_serialize::{json, Encodable, Encoder};
 use error::{Result, Error};
-use payload::Payload;
+use {Payload, SlackText, serde_json};
+use serde::{Serialize, Serializer};
 
 /// Handles sending messages to slack
 #[derive(Debug)]
@@ -21,7 +21,7 @@ impl Slack {
     /// Send payload to slack service
     pub fn send(&self, payload: &Payload) -> Result<()> {
         debug!("sending payload, {:?}", payload);
-        let encoded = try!(json::encode(payload));
+        let encoded = try!(serde_json::to_string(payload));
         debug!("JSON payload, {:?}", encoded);
         let mut easy = Easy::new();
         let _ = easy.url(&self.incoming_url[..]);
@@ -49,11 +49,6 @@ impl Slack {
         }
     }
 }
-
-/// Representation of any text sent through slack
-/// the text must be processed to escape specific characters
-#[derive(Debug, Default, Clone)]
-pub struct SlackText(String);
 
 impl SlackText {
     /// Construct slack text with escaping
@@ -117,13 +112,6 @@ impl ::std::fmt::Display for SlackText {
     }
 }
 
-impl Encodable for SlackText {
-    fn encode<S: Encoder>(&self, encoder: &mut S) -> ::std::result::Result<(), S::Error> {
-        let text = format!("{}", &self);
-        encoder.emit_str(&text)
-    }
-}
-
 /// Representation of a link sent in slack
 #[derive(Debug)]
 pub struct SlackLink {
@@ -149,10 +137,11 @@ impl ::std::fmt::Display for SlackLink {
     }
 }
 
-impl Encodable for SlackLink {
-    fn encode<S: Encoder>(&self, encoder: &mut S) -> ::std::result::Result<(), S::Error> {
-        let text = format!("{}", &self);
-        encoder.emit_str(&text)
+impl Serialize for SlackLink {
+    fn serialize<S>(&self, serializer: &mut S) -> ::std::result::Result<(), S::Error>
+        where S: Serializer
+    {
+        serializer.serialize_str(&format!("{}", self)[..])
     }
 }
 
@@ -160,9 +149,8 @@ impl Encodable for SlackLink {
 mod test {
     #[cfg(feature = "unstable")]
     use test::Bencher;
-    use slack::{Slack, SlackLink, SlackText};
-    use {PayloadBuilder, AttachmentBuilder, Field};
-    use rustc_serialize::json;
+    use slack::{Slack, SlackLink};
+    use {PayloadBuilder, AttachmentBuilder, Field, SlackText, serde_json};
 
     #[test]
     fn slack_incoming_url_test() {
@@ -193,7 +181,7 @@ mod test {
             text: SlackText::new("moo <&> moo"),
             url: "http://google.com".to_owned(),
         };
-        assert_eq!(json::encode(&s).unwrap().to_owned(),
+        assert_eq!(serde_json::to_string(&s).unwrap().to_owned(),
                    "\"<http://google.com|moo &lt;&amp;&gt; moo>\"".to_owned())
     }
 
@@ -213,18 +201,19 @@ mod test {
             .icon_emoji(":chart_with_upwards_trend:")
             .attachments(a)
             .unfurl_links(false)
-            .link_names(false)
+            .link_names(true)
             .build()
             .unwrap();
 
-        assert_eq!(json::encode(&p).unwrap().to_owned(), r##"{"text":"test message","channel":"#abc","username":"Bot","icon_url":null,"icon_emoji":":chart_with_upwards_trend:","attachments":[{"fallback":"fallback &lt;&amp;&gt;","text":"text &lt;&amp;&gt;","pretext":null,"color":"#6800e8","fields":[{"title":"title","value":"value","short":null}]}],"unfurl_links":false,"unfurl_media":null,"link_names":0}"##.to_owned())
+        assert_eq!(serde_json::to_string(&p).unwrap().to_owned(), r##"{"text":"test message","channel":"#abc","username":"Bot","icon_emoji":":chart_with_upwards_trend:","attachments":[{"fallback":"fallback &lt;&amp;&gt;","text":"text &lt;&amp;&gt;","color":"#6800e8","fields":[{"title":"title","value":"value"}]}],"unfurl_links":false,"link_names":1}"##.to_owned())
     }
 
     #[test]
     fn json_message_payload_test() {
         let p = PayloadBuilder::new().text("test message").build().unwrap();
 
-        assert_eq!(json::encode(&p).unwrap().to_owned(), r##"{"text":"test message","channel":null,"username":null,"icon_url":null,"icon_emoji":null,"attachments":null,"unfurl_links":null,"unfurl_media":null,"link_names":null}"##.to_owned())
+        assert_eq!(serde_json::to_string(&p).unwrap().to_owned(),
+                   r##"{"text":"test message"}"##.to_owned())
     }
 
     #[test]